aws-lambda-s3-thumbnail/requirement.txt
appdirs==1.4.3
boto3==1.4.4
botocore==1.5.30
docutils==0.13.1
futures==3.0.5
jmespath==0.9.2
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.2.0
python-dateutil==2.6.0
requests==2.13.0
python-resize-image==1.1.11
s3transfer==0.1.10
six==1.10.0